+++ Model/Layers/Layer.cs
@@ -2,12 +2,12 @@
 
 public abstract class Layer
 {
-    public double[,] Inputs { get; set; }
-    public double[,] Output { get; set; }
-    public double[,] Dinputs { get; set; }
+    public float[,] Inputs { get; set; }
+    public float[,] Output { get; set; }
+    public float[,] Dinputs { get; set; }
     public Layer Prev { get; set; }
     public Layer Next { get; set; }
 
-    public abstract void Forward(double[,] inputs);
-    public abstract void Backward(double[,] dvalues);
+    public abstract void Forward(float[,] inputs);
+    public abstract void Backward(float[,] dvalues);
 }
